created state.ts
